rewrite YarnAppSecurityService to wxpose an interface to make the security manager be configurable and can be reimplemented
